feat: move theme change from app to header
